api/requirements.txt
flask
flask-cors
google-cloud-firestore
beautifulsoup4 
requests 
google-cloud-storage
vertexai